HANG-MAN CODE Using Jack
hangman. noun. hang·​man ˈhaŋ-mən. : a person who hangs condemned criminals. : a game in which the object is for one player to guess the letters of an unknown word before the player who knows the word creates a stick figure of a hanged man by drawing one line for each incorrect guess.